tests/submission/Infraware/Forms/contents/Forms/Not Use/textarea_rows.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>textarea_rows</h3>
	</p>
	
	<li>Rows of textarea control is 3.</li>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form">
	<p><textarea id='textarea_id'></textarea></p>
	</form>

	<script>
		
		var textarea = document.getElementById("textarea_id");
		textarea.rows = 3;

		if (typeof(textarea.rows) == "number") {
			if (textarea.rows != 3) {
				test( function() {
					assert_true(false, "rows attribute is not correct.");
				});
			}
		} else {
			test( function() {
				assert_unreached("rows attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
